Changes to remove warnings in MSVC build.
FossilOrigin-Name: 6cf76c2ae25d6e58926637ecd42eed6b300b1a25

+++ ext/fts3/fts3.c
@@ -2210,8 +2210,9 @@ static void fts3ExprMatchInfo(
             sqlite3_int64 iOffset = 0;
             char *pList = pExpr->pCurrent;
             while( *pList&0xFE ){
-              fts3GetDeltaVarint(&pList, &iOffset); iOffset -= 2;
-              fts3MatchInfoAppend(pInfo, iOffset+1-nPhrase);
+              fts3GetDeltaVarint(&pList, &iOffset);
+              iOffset -= 2;
+              fts3MatchInfoAppend(pInfo, (unsigned int)(iOffset+1-nPhrase));
               nLocal++;
             }
             pExpr->pCurrent = pList;
